Convert run-computation to file-based router
