+++ strudel-demo-app/src/pages/compare-data/_config/definitions.json
