+++ strudel-demo-app/src/pages/run-computation/[id]/data-inputs.tsx
@@ -1,52 +1,30 @@
-import { Box, Button, Container, Grid, Link, Paper, Stack, Step, StepLabel, Stepper, Typography } from '@mui/material';
-import { DataGrid } from '@mui/x-data-grid';
+import { Box, Button, Container, Link, Paper, Stack, Step, StepLabel, Stepper, Typography } from '@mui/material';
 import React, { useEffect } from 'react';
-import Plot from 'react-plotly.js';
 import { Link as RouterLink } from 'react-router-dom';
-import { useDataFromSource } from '../../utils/useDataFromSource';
-import { basename } from '../../main';
-import { useRunComputation } from './_context/ContextProvider';
-import { setResultsBarChartData, setResultsLineChartData, setResultsTableData } from './_context/actions';
+import { DataGrid } from '@mui/x-data-grid';
+import { useDataFromSource } from '../../../utils/useDataFromSource';
+import { basename } from '../../../main';
+import { useRunComputation } from '../_context/ContextProvider';
+import { setInputsTableData } from '../_context/actions';
 
 /**
- * Results page to display after a computation completes in the run-computation Task Flow.
- * Displays a line chart, bar chart, and table of results from the computation.
+ * Page to display input data after creating or selecting an item from 
+ * the `<ComputationsList>` page in the run-computation Task Flow.
+ * Table columns are configured in `definitions.inputs.table.columns`
  */
-export const Results: React.FC = () => {
+const DataInputsPage: React.FC = () => {
   const { state, dispatch } = useRunComputation();
   // strudel-kit-variable-next-line
-  const tableData = useDataFromSource('default/run-computation/results_table.json', basename);
-  // strudel-kit-variable-next-line
-  const lineData = useDataFromSource('default/run-computation/results_line_chart.json', basename);
-  // strudel-kit-variable-next-line
-  const barData = useDataFromSource('default/run-computation/results_bar_chart.json', basename);
-
+  const inputsData = useDataFromSource('default/run-computation/inputs.json', basename);
+  
   /**
-   * Set data for the results table when the data loads
+   * Set data for the inputs table when the data loads
    */
   useEffect(() => {
-    if (!state.results.table.data || state.results.table.data.length === 0) {
-      dispatch(setResultsTableData(tableData));
+    if (!state.inputs.table.data || state.inputs.table.data.length === 0) {
+      dispatch(setInputsTableData(inputsData));
     }
-  }, [tableData]);
-
-  /**
-   * Set data for the results line chart when the data loads
-   */
-  useEffect(() => {
-    if (!state.results.lineChart.data || state.results.lineChart.data.length === 0) {
-      dispatch(setResultsLineChartData(lineData));
-    }
-  }, [lineData]);
-
-  /**
-   * Set data for the results bar chart when the data loads
-   */
-  useEffect(() => {
-    if (!state.results.barChart.data || state.results.barChart.data.length === 0) {
-      dispatch(setResultsBarChartData(barData));
-    }
-  }, [barData]);
+  }, [inputsData]);
 
   /**
    * Content to render on the page for this component
@@ -61,11 +39,11 @@ export const Results: React.FC = () => {
           borderColor: 'neutral.main'
         }}
       >
-        <Stepper activeStep={2} sx={{ maxWidth: 850 }}>
+        <Stepper activeStep={0} sx={{ maxWidth: 850 }}>
           <Step key="Data Inputs">
             <StepLabel>
               <Link component={RouterLink} to="../data-inputs" sx={{ color: 'inherit', textDecoration: 'none' }}>
-                {/* strudel-kit-variable-next-line   */}
+                {/* strudel-kit-variable-next-line */}
                 Data Inputs
               </Link>
             </StepLabel>
@@ -121,7 +99,7 @@ export const Results: React.FC = () => {
               marginRight: '-2rem !important'
             }}
           >
-            Summary
+            Input Units
           </Typography>
           <Typography 
             component="li"
@@ -131,7 +109,7 @@ export const Results: React.FC = () => {
               marginRight: '-2rem !important'
             }}
           >
-            System Costing
+            Input Streams
           </Typography>
           <Typography 
             component="li" 
@@ -141,45 +119,25 @@ export const Results: React.FC = () => {
               marginRight: '-2rem !important'
             }}
           >
-            System Metrics
+            Unit Costing
           </Typography>
         </Stack>
-        <Box flex={1}>
+        <Box flex={1} sx={{ overflow: 'hidden' }}>
           <Container
             maxWidth="xl"
             sx={{
               mt: 4
             }}
           >
-            <Grid container spacing={4}>
-              <Grid item sm={6}>
-                <Paper>
-                  <Plot
-                    data={state.results.lineChart.data}
-                    layout={{}}
-                  />
-                </Paper>
-              </Grid>
-              <Grid item sm={6}>
-                <Paper>
-                  <Plot
-                    data={state.results.barChart.data}
-                    layout={{}}
-                  />
-                </Paper>
-              </Grid>
-              <Grid item xs={12}>
-                <Paper>
-                  <DataGrid
-                    rows={state.results.table.data || []}
-                    getRowId={(row) => row[state.results.table.dataIdField]}
-                    columns={state.results.table.columns}
-                    disableColumnSelector
-                    disableRowSelectionOnClick
-                  />
-                </Paper>
-              </Grid>
-            </Grid>  
+            <Paper>
+              <DataGrid
+                rows={state.inputs.table.data || []}
+                getRowId={(row) => row[state.inputs.table.dataIdField]}
+                columns={state.inputs.table.columns}
+                disableColumnSelector
+                disableRowSelectionOnClick
+              />
+            </Paper>  
           </Container>
         </Box>
       </Stack>
@@ -191,14 +149,17 @@ export const Results: React.FC = () => {
           bottom: 0,
           padding: 2,
           position: 'fixed',
+          textAlign: 'right',
           width: '100%'
         }}
       >
         <Link component={RouterLink} to="../settings">
           {/* strudel-kit-variable-next-line */}
-          <Button variant="contained">Back to optimization settings</Button>
+          <Button variant="contained">Continue to optimization settings</Button>
         </Link>
       </Box>
     </Stack>
   )
 }
+
+export default DataInputsPage;
