+++ strudel-demo-app/src/pages/run-computation/_config/definitions.json
